+++ src/apify/__init__.py
@@ -1,6 +1,9 @@
-from ._version import __version__
+from importlib import metadata
+
 from .actor import Actor
 from .config import Configuration
 from .proxy_configuration import ProxyConfiguration, ProxyInfo
+
+__version__ = metadata.version('apify')
 
 __all__ = ['Actor', 'Configuration', 'ProxyConfiguration', 'ProxyInfo', '__version__']